Modernize the Build
