+++ streamlit_label_kit/__init__.py
@@ -1,5 +1,13 @@
-IS_RELEASE = True
+#
+# Streamlit components for general labeling tasks
+#
+# Copyright (c) 2024 Carnegie Mellon University
+# SPDX-License-Identifier: GPL-2.0-only
+#
 
-from .LabelToolKit.detection import detection, convert_bbox_format, absolute_to_relative, relative_to_absolute, are_bboxes_equal
+from .LabelToolKit import convert_bbox_format, absolute_to_relative, relative_to_absolute
+from .LabelToolKit.detection import detection
 from .LabelToolKit.annotation import annotation
 from .LabelToolKit.segmentation import segmentation
+
+__version__ = "0.1.0"
